services/booking-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
kafka-python
